Adding an event system where event subscribers are notified when an event publisher calls deliver. Events may carry templated data that notified subscribers may access when they are notified.

+++ source/Library.Shared/ActionIncrement.h
@@ -90,7 +90,7 @@ namespace FIEAGameEngine
 		/// <summary>
 		/// 
 		/// </summary>
-		float mStep = 1;
+		int mStep = 1;
 	};
 
 	CONCRETE_FACTORY(ActionIncrement, Scope);
